Add new lines for readability, rename command file
