+++ src/commands/synthetics/cli.ts
@@ -1,4 +1,4 @@
 import {RunTestsCommand} from './run-tests-command'
-import {UploadApplicationCommand} from './uploadApplicationCommand'
+import {UploadApplicationCommand} from './upload-application-command'
 
 module.exports = [RunTestsCommand, UploadApplicationCommand]
