+++ src/commands/synthetics/upload-application-command.ts
